refactor: extract shared SettingsModelSelect to eliminate duplication
