+++ apps/web/src/components/settings/model-select.tsx
@@ -1,0 +1,131 @@
+import * as React from 'react';
+
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import {
+  Combobox,
+  ComboboxCollection,
+  ComboboxContent,
+  ComboboxEmpty,
+  ComboboxGroup,
+  ComboboxInput,
+  ComboboxItem,
+  ComboboxLabel,
+  ComboboxList,
+  ComboboxSeparator,
+} from '@/components/ui/combobox';
+import {
+  deleteSettingMutationOptions,
+  saveSettingMutationOptions,
+} from '@/lib/queries/settings';
+import type { ProviderModels } from '@/lib/queries/providers';
+
+export type ModelOption = {
+  label: string;
+  providerId: string;
+  modelId: string;
+};
+
+type ModelGroup = {
+  value: string;
+  items: ModelOption[];
+};
+
+export function buildGroupedItems(providerModels: ProviderModels[]): ModelGroup[] {
+  return providerModels.map((provider) => ({
+    value: provider.providerName,
+    items: provider.models.map((model) => ({
+      label: model.name,
+      providerId: provider.providerId,
+      modelId: model.id,
+    })),
+  }));
+}
+
+export function flattenGroups(groups: ModelGroup[]): ModelOption[] {
+  return groups.flatMap((g) => g.items);
+}
+
+type SettingsModelSelectProps = {
+  providerIdKey: string;
+  modelIdKey: string;
+  currentProviderId: string | undefined;
+  currentModelId: string | undefined;
+  providerModels: ProviderModels[];
+  placeholder?: string;
+  showClear?: boolean;
+};
+
+export function SettingsModelSelect({
+  providerIdKey,
+  modelIdKey,
+  currentProviderId,
+  currentModelId,
+  providerModels,
+  placeholder = 'Search models...',
+  showClear,
+}: SettingsModelSelectProps) {
+  const queryClient = useQueryClient();
+
+  const groups = React.useMemo(() => buildGroupedItems(providerModels), [providerModels]);
+  const allOptions = React.useMemo(() => flattenGroups(groups), [groups]);
+
+  const saveProviderMutation = useMutation(saveSettingMutationOptions(providerIdKey, queryClient));
+  const saveModelMutation = useMutation(
+    saveSettingMutationOptions(modelIdKey, queryClient, { silent: true }),
+  );
+  const deleteProviderMutation = useMutation(
+    deleteSettingMutationOptions(providerIdKey, queryClient),
+  );
+  const deleteModelMutation = useMutation(
+    deleteSettingMutationOptions(modelIdKey, queryClient, { silent: true }),
+  );
+
+  function handleValueChange(value: ModelOption | null) {
+    if (!value) {
+      if (currentProviderId) deleteProviderMutation.mutate();
+      if (currentModelId) deleteModelMutation.mutate();
+      return;
+    }
+    saveProviderMutation.mutate(value.providerId);
+    saveModelMutation.mutate(value.modelId);
+  }
+
+  const selectedOption =
+    currentProviderId && currentModelId
+      ? (allOptions.find(
+          (o) => o.providerId === currentProviderId && o.modelId === currentModelId,
+        ) ?? null)
+      : null;
+
+  const resolvedShowClear = showClear ?? !!(currentProviderId && currentModelId);
+
+  return (
+    <Combobox<ModelOption>
+      value={selectedOption}
+      onValueChange={handleValueChange}
+      isItemEqualToValue={(a, b) => a.providerId === b.providerId && a.modelId === b.modelId}
+      items={groups}
+    >
+      <ComboboxInput placeholder={placeholder} showClear={resolvedShowClear} />
+      <ComboboxContent side="bottom" sideOffset={4} align="start">
+        <ComboboxEmpty>No models found</ComboboxEmpty>
+        <ComboboxList>
+          {(group, index) => (
+            <ComboboxGroup key={group.value} items={group.items}>
+              <ComboboxLabel>{group.value}</ComboboxLabel>
+              <ComboboxCollection>
+                {(item) => (
+                  <ComboboxItem key={`${item.providerId}:${item.modelId}`} value={item}>
+                    {item.label}
+                  </ComboboxItem>
+                )}
+              </ComboboxCollection>
+              {index < groups.length - 1 && <ComboboxSeparator />}
+            </ComboboxGroup>
+          )}
+        </ComboboxList>
+      </ComboboxContent>
+    </Combobox>
+  );
+}
